(REFACTOR) presentation util 파일 구조 정리 - MapLogicHandler를 MapUtils에 병합, bottomShadow.kt를 ModifierExtensions.kt로 이름 변경 (#20)
